pyproject.toml
[tool.poetry]
name = "victory"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.89.1"
uvicorn = { version = "^0.20.0", extras = ["standard"] }
pydantic = {version = "^1.10.4", extras = ["dotenv"]}
yarl = "^1.8.2"
ujson = "^5.7.0"
aiofiles = "^22.1.0"
httptools = "^0.5.0"
loguru = "^0.6.0"
yoyo-migrations = "^8.2.0"
psycopg2-binary = "^2.9.6"
cursor = "^1.3.5"
aiopg = "^1.4.0"
bcrypt = "^4.0.1"
fastapi-jwt = "^0.1.8"
hiredis = "^2.2.3"


[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
fakeredis = "^2.5.0"
httpx = "^0.23.3"
taskiq = { version = "^0", extras = ["reload"] }

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["victory",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

# Remove this and add `types-redis`
# when the issue https://github.com/python/typeshed/issues/8242 is resolved.
[[tool.mypy.overrides]]
module = [
    'redis.asyncio'
]
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "VICTORY_ENVIRONMENT=pytest",
    "VICTORY_DB_BASE=victory_test",
]

[fastapi-template.options]
project_name = "victory"
api_type = "rest"
enable_redis = "True"
enable_rmq = "None"
ci_type = "none"
enable_migrations = "True"
enable_taskiq = "None"
enable_kube = "None"
kube_name = "victory"
enable_routers = "True"
enable_kafka = "None"
enable_loguru = "True"
traefik_labels = "None"
add_dummy = "True"
orm = "none"
self_hosted_swagger = "True"
prometheus_enabled = "None"
sentry_enabled = "None"
otlp_enabled = "None"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
